A bit of refactoring
